experiments/vision-roundtrip/requirements.txt
torch>=2.1
transformers>=4.40
Pillow>=10.0
numpy>=1.24
einops>=0.7
accelerate>=0.27
